+++ profiles_api/views.py
@@ -14,6 +14,8 @@ from profiles_api import serializers
 from profiles_api import models
 from profiles_api import permissions
 
+from profiles_api.topic.topic_model import Topic
+
 
 
 class HelloApiView(APIView):
@@ -166,18 +168,6 @@ class QuestionViewSet(viewsets.ModelViewSet):
         return questions
 
 
-class TopicViewSet(viewsets.ModelViewSet):
-    """Handles creating, reading and updating topics"""
-    authentication_classes = (TokenAuthentication,)
-    serializer_class = serializers.TopicSerializer
-    queryset = models.Topic.objects.all()
-    permission_classes = (permissions.UpdateOwnStatus, IsAuthenticated)
-
-    def perform_create(self, serializer):
-        """Sets the user profile to the logged in user"""
-        serializer.save(user_profile=self.request.user)
-
-
 class SubtopicViewSet(viewsets.ModelViewSet):
     """Handles creating, reading and updating subtopics"""
     authentication_classes = (TokenAuthentication,)
@@ -325,9 +315,9 @@ class QuestionView(APIView):
                 return Response(data='Topic not defined', status=400)
             else:
                 filter_dict = {'id': topic_id}
-                topic = models.Topic.objects.filter(**filter_dict)[0]
+                topic = Topic.objects.filter(**filter_dict)[0]
         else:
-            topic = models.Topic.objects.get_or_create(
+            topic = Topic.objects.get_or_create(
                 name=topic_name,
                 user_profile_id=user.id
             )[0]
@@ -437,7 +427,7 @@ class CustomSubtopicView(APIView):
         if user is None or user.id == '':
             return Response(data='User not defined', status=400)
 
-        topic = models.Topic.objects.get_or_create(
+        topic = Topic.objects.get_or_create(
             name=topic_name,
             user_profile_id=user.id
         )[0]
@@ -451,50 +441,6 @@ class CustomSubtopicView(APIView):
 
         serializer = serializers.SubTopicSerializer(subtopic)
 
-        return Response(data=serializer.data, status=201)
-
-
-class TopicView(APIView):
-    """Custom view for topics"""
-    authentication_classes = (TokenAuthentication,)
-    permission_classes = (permissions.UpdateOwnStatus, IsAuthenticated)
-
-    def get(self, request):
-        """Retrieve only certain topics"""
-        start = self.request.query_params.get('start', None)
-        number = self.request.query_params.get('number', None)
-
-        topics = models.Topic.objects.all()
-
-        if start is not None:
-            topics = topics[min(abs(int(start)), topics.count()):]
-
-        if number is not None:
-            topics = topics[:max(0, min(int(number), topics.count()))]
-
-        if topics.count() > 0:
-            serializer = serializers.TopicSerializer(topics, many=True)
-            return Response(data=serializer.data, status=200)
-        else:
-            return Response(status=204)
-
-    def post(self, request):
-        """Create a new topic"""
-        topic_name = request.data['topic']
-        if topic_name is None or topic_name == '':
-            return Response(data='Topic not defined', status=400)
-
-        user = self.request.user
-
-        if user is None or user.id == '':
-            return Response(data='User not defined', status=400)
-        else:
-            topic = models.Topic.objects.get_or_create(
-                name=topic_name,
-                user_profile_id=user.id
-            )[0]
-
-        serializer = serializers.TopicSerializer(topic)
         return Response(data=serializer.data, status=201)
